+++ docker/pytorch-aarch64/CHANGELOG.md
@@ -15,18 +15,31 @@ where `YY` is the year, and `MM` the month of the increment.
 
 ### Fixed
 
+## [r22.05] 2022-05-06
+https://github.com/ARM-software/Tool-Solutions/tree/tensorflow-pytorch-aarch64--r22.05/docker/pytorch-aarch64
+
+### Added
+- Python vision examples updated to allow use of a local image file.
+
+### Changed
+
+### Removed
+
+### Fixed
+- Python vision examples will not download labels and images if they are already present.
+
 ## [r22.04] 2022-04-08
 https://github.com/ARM-software/Tool-Solutions/tree/tensorflow-pytorch-aarch64--r22.04/docker/pytorch-aarch64
 
 ### Added
-- Adds docker/pytorch-aarch64/CHANGELOG.md
+- Adds docker/pytorch-aarch64/CHANGELOG.md.
 
 ### Changed
-- Updates oneDNN to v2.6
-- Does not download model in example scripts if it was downloaded already
+- Updates oneDNN to v2.6.
+- Does not download model in example scripts if it was downloaded already.
 
 ### Fixed
-- Fixes git submodule init in SciPy build
+- Fixes git submodule init in SciPy build.
 
 ## [r22.03] 2022-03-18
 https://github.com/ARM-software/Tool-Solutions/tree/tensorflow-pytorch-aarch64--r22.03/docker/pytorch-aarch64